+++ CHANGELOG.md
@@ -18,3 +18,9 @@
 ### Fix
 
 - path export image
+
+### Changed
+
+- update lib to work with husky
+- update from next 14 to 15
+- update from modalName and drawerName to id